GlobalVariablesOverride: bow out earlier if a whitelist comment is found
If the token being examined is a `T_VARIABLE` token and has a whitelist comment, the sniff can bow out before examining the token in detail.

Includes splitting the `maybe_add_error()` method up to prevent checking twice for a whitelist comment.

For the `T_GLOBAL` token this is not possible as the whitelist comment will not be on the line with the `$stackPtr` pointing to the `global` keyword.