GlobalVariablesOverride: add `treat_files_as_scoped` property
By setting the `treat_files_as_scoped` property to `true`, the sniff will regard each file as being included from with a function, making it scoped to the function holding the `include` statement.

This basically allows for the sniff to revert back to the original behaviour of only examining variables in the global scope if a `global` statement had been found or if the variable used was `$GLOBALS`.