Docs: Remove and update VIP link references
The WordPress.com VIP team moved their coding standards documentation around, since they now have different documents for their different platforms:

- [WPCOM platform](https://lobby.vip.wordpress.com/wordpress-com-documentation/code-review-what-we-look-for/) (only available to those with VIP Lobby access)
- [VIP-Go platform](https://vip.wordpress.com/documentation/vip-go/code-review-blockers-warnings-notices/)

Many of the existing link references were out of date and pointed to pages or page anchors that didn't exist. Since our `WordPress-VIP` standard is deprecated, and will be removed in `2.0.0`, it's easier to remove all of the link references in the `VIP` category ruleset and sniffs.

The only ones that remain are ones that are used as references for sniffs typically found in `WordPress-Extra` (outside of the `VIP` category). They have been updated as necessary.