PrecisionAlignment: bug fix
This issue was discovered through a report upstream in issue squizlabs/PHP_CodeSniffer 2179#issuecomment-437634372.

What it comes down to is that if the `ignoreAlignmentTokens` property would be set, the `merge_custom_array()` call in the `process_token()` would override it with a flipped array with values `false` and for the next file, would override it again, flipping the array again etc etc.

This in effect made the custom property only work for the first file, not for subsequent files.

This has now been fixed.

Includes unit tests to guard against the issue.

Additionally, I've improved the inline comments to clarify what's being tested in test case file 2.