PHPUnit: add Composer script to run the tests
A `dev` requirement of `phpunit/phpunit` has been added to the `composer.json` file, as well as a script to run the unit tests.
While PHPUnit is not strictly speaking a dependency of WPCS (but of PHPCS), this just makes life easier on WPCS developers who use the Composer install for development.

If - as a WPCS sniff developer - you've installed WPCS using Composer, you can now run the unit tests, like so:
```bash
composer run-tests
```

A note to this effect has been added to the `CONTRIBUTING.MD` file as well.

Additionally:
* While not strictly necessary, a `testsuite` directive has been added to the `phpunit.xml.dist` file to document where the tests are located and how they are named.
* As `PHPCompatibility` will also be installed when using Composer to install WPCS and PHPCompatibility uses its own test setup, we need to prevent PHPCS from trying to load the test files for PHPCompatibility.
    The `PHPCS_IGNORE_TEST` environment variable in the `phpunit.xml.dist` file does just that.
    For more information about this, see squizlabs/PHP_CodeSniffer 1146