Rulesets: minor simplifications (PHPCS 3.0+)
As of PHPCS 3.0.0, you can include individual error codes, not just complete sniffs.

Where WPCS rulesets included sniffs, but excluded more than 50% of the error codes from that sniff, it makes sense to switch them over to include individual error codes instead.

This also means that where previously, if the `type` and/or `message` of an individual error code was adjusted via the ruleset, the sniff needed to be included first and only after that, the details could be adjusted.
Now, the inclusion can just be done based on error code, providing all error codes in the sniff are adjusted.

To explain it better:
* If a sniff contained five error codes and we adjusted the message for one, there is no real advantage to changing over.
* If a sniff contained three error codes and we excluded one, and adjusted another, there is no real advantage to changing over.
* However, if a sniff contained only one error code and we'd adjust that one anyway, the new way of including just the error code gets rid of some superfluous configuration.
* Similarly, if a sniff contained five error codes, we'd exclude two and adjusted the details for the other three, changing over, again, is to our benefit.