Core: Defer to upstream `PSR12.Keywords.ShortFormTypeKeywords` sniff (PHPCS 3.3.0)
The WPCS native `WordPress.PHP.TypeCasts` sniff used to check that:
* Type casts are in lowercase;
* Boolean and integer casts are short form;
* The `(double)` cast is not used;
* Discourages the use of the `(unset)` and `(binary)` casts.

The lowercase checks has been removed in WPCS 2.0.0 in favour of the upstream `Generic.PHP.LowerCaseType` sniff via 1582.

Asides from that sniff, PHPCS 3.3.0 also introduces the `PSR12.Keywords.ShortFormTypeKeywords` sniff which checks that boolean and integer casts are in short form.

This PR removes that check from the `WordPress.PHP.TypeCasts` sniff in favour of using the upstream sniff.

There is currently one known bug in the upstream sniff - a false positive when a type cast contains whitespace, see squizlabs/PHP_CodeSniffer 2331 -, however, as WPCS also includes the `Squiz.WhiteSpace.CastSpacing` sniff which forbids whitespace within type casts anyway, this is not a reason to delay.

Includes updating the ruleset inline documentation to be in line with the current text in the handbook.
Includes updating the sniff class documentation to be in line with the current checks contained in the sniff.
Includes removing unit tests which are no longer relevant.