Sniff: add two new utility methods `is_class_object_call()` and `is_token_namespaced()`
These should help make the checking of whether or not a function call is a global function, method call or a namespaced function call more consistent.

These functions will be unit tested by the implementations in various other places and the fact that the existing unit tests still pass after that.