Sniff::has_nonce_check(): ignore nonce checks in nested closed scopes
A function, closure, anonymous class and other OO constructs are all "closed" scopes.

If any of these are nested, they open a nested closed scope and anything within that scope should be disregarded for the purpose of verifying whether or not a nonce check has been executed.

This small change implements that.

This fixes some potential false negatives.

Includes unit tests.

Related to 764