Sniff::is_test_class(): recognize `TestCase`
What with WP dropping support for PHP 5.2, I expect that more and more projects will start to use `use` statements in their code.

When the PHPUnit native `PHPUnit\Framework\TestCase` is imported via a `use` statement and then `extend`ed, the WPCS sniffs would not recognize the extending class as a test class.

Adding `TestCase` to the default test class whitelist fixes this.

While this means that any class extending the quite genericly named `TestCase` will now be recognized as a test class, even when the `TestCase` in question is not the PHPUnit native one, I think it's safe to assume that even when it's not the PHPUnit native one, it _will_ be a test base class and for WPCS purposes the extending class can be marked as a test class.

Includes unit test via the `GlobalVariablesOverride` sniff.

Includes adjusting an existing unit test for the `FileName` sniff to still test what it was testing before.