Sniff::is_short_list(): switch over to a PHPCSUtils alternative
PHPCSUtils contains both a `Lists::isShortList()` as well as an `Arrays::isShortArray()` method. These are not direct opposites as there are tokenizer problems across PHPCS version which sometimes affect one, sometimes the other, so instead of checking that something is "not a short list", in these cases, we should be checking that something **is** a short array.