EscapeOutput: defensive coding
The `PassedParameters::getParameter()` method (and the `Sniff::get_function_call_parameter()` method before it), either returns an array with information on the parameter or `false` if the parameter does not exist.

A call to `trigger_error()` and the likes, without passing parameters is a little useless and will throw a warning in PHP 4/5/7 and a fatal `ArgumentCountError` in PHP 8.

However, those types of errors are not the concern of this sniff, so if a call to these functions is encountered without arguments, the sniff should conclude that there is nothing which needs escaping and bow out.

Fix now, including unit tests to safeguard against it in the future.