PHP 8.0 | DeprecationHelper: allow for attributes between function declaration and docblock
PHP 8.0 introduced attributes, which can be placed - and commonly are placed - between a function declaration and its docblock.
Ref: https://www.php.net/manual/en/language.attributes.php

This commit updates the `is_function_deprecated()` method to skip over attributes. This prevents false positives when a function is deprecated, but also has an attribute.

Tested via both sniffs using the helper class.

Includes renaming a local variable to be more descriptive (`$ignore` instead of `$find`).