Move "is comparison" related utility to dedicated `VariableHelper`
The "is comparison" related utility method is only used by a small set of sniffs, so is better placed in a dedicated class.

As this method also expects a `T_VARIABLE` token as input, the `VariableHelper` class seems appropriate.

This commit moves the `is_comparison()` method to the new `WordPressCS\WordPress\Helpers\VariableHelper` and starts using that class in the relevant sniffs.

**Note**:
It is expected for PHPCSUtils to have dedicated methods for the same at some point in the future. If/when those methods become available, it is recommended for the sniffs to start using the PHPCSUtils methods.

Related to 1465