NoSilencedErrors: rename property [2]
As this is a `public` property, this is a BC-break and should be annotated as such in the upgrade guide/changelog.

Also note that the new name is in line with the name for the same property in an upcoming replacement sniff in PHPCSExtra. This should make the switch over to the replacement sniff more straight-forward (if and when).